backend/requirements.txt
blinker==1.9.0
click==8.2.1
config==0.5.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.1
Werkzeug==3.1.3
faiss-cpu==1.8.0
numpy==1.24.3
sentence-transformers==2.2.2
